First refactoring step: Move functions related to events model and list model to separate modules
