+++ src/lib.rs
@@ -15,6 +15,6 @@
 mod item_sort_list;
 pub mod main_window;
 mod misc;
+mod models;
 mod persistence;
-mod gui_items;
 mod synchronize;
