+++ src/models/mod.rs
@@ -1,0 +1,3 @@
+pub mod events_model;
+pub mod gui_items;
+pub mod list_model;
